# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creators Dashboard profile API
  description: |
    A Flask + MongoDB API for the profile service.
    
    ## Authentication
    All endpoints (except `/dev-login` and `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a token from the `/dev-login` endpoint.
  version: 0.1.0
  contact:
    name: Agile Crafts People
    email: info@agilecrafts.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:9096
    description: Local development server
...
paths:
...
  /api/Platform:
    get:
      tags:
        - Platform
      summary: Get all Platform documents
      description: Retrieve infinite scroll batch of sorted, filtered Platform documents
      operationId: getPlatforms
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Platform
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Platform
      summary: Create a new Platform document
      description: Create a new Platform document (requires staff or admin role)
      operationId: createPlatform
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformInput'
      responses:
        '201':
          description: Platform created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Platform/{PlatformId}:
    get:
      tags:
        - Platform
      summary: Get Platform by PlatformId
      description: Retrieve a specific Platform document by its PlatformId
      operationId: getPlatform
      security:
        - bearerAuth: []
      parameters:
        - name: PlatformId
          in: path
          required: true
          description: The Platform ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Platform
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Platform
      summary: Update Platform document
      description: Update a Platform document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updatePlatform
      security:
        - bearerAuth: []
      parameters:
        - name: PlatformId
          in: path
          required: true
          description: The Platform ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformUpdate'
      responses:
        '200':
          description: Platform updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/User:
    get:
      tags:
        - User
      summary: Get all User documents
      description: Retrieve infinite scroll batch of sorted, filtered User documents
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-User
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - User
      summary: Create a new User document
      description: Create a new User document (requires staff or admin role)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/User/{UserId}:
    get:
      tags:
        - User
      summary: Get User by UserId
      description: Retrieve a specific User document by its UserId
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: UserId
          in: path
          required: true
          description: The User ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - User
      summary: Update User document
      description: Update a User document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: UserId
          in: path
          required: true
          description: The User ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/dev-login` endpoint.
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
...
    Platform:
      type: object
      description: Platform domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Platform name
          pattern: '^[^\s]{1,40}$'
          example: my-Platform
        description:
          type: string
          description: Platform description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Platform document
        status:
          type: string
          description: Platform status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    PlatformInput:
      type: object
      description: Input for creating a Platform document
      properties:
        name:
          type: string
          description: Platform name
          pattern: '^[^\s]{1,40}$'
          example: my-Platform
        description:
          type: string
          description: Platform description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Platform document
        status:
          type: string
          description: Platform status
          enum:
            - active
            - archived
          example: active
    PlatformUpdate:
      type: object
      description: Fields to update in a Platform document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Platform name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Platform description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Platform status
          enum:
            - active
            - archived
    User:
      type: object
      description: User domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: User name
          pattern: '^[^\s]{1,40}$'
          example: my-User
        description:
          type: string
          description: User description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample User document
        status:
          type: string
          description: User status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    UserInput:
      type: object
      description: Input for creating a User document
      properties:
        name:
          type: string
          description: User name
          pattern: '^[^\s]{1,40}$'
          example: my-User
        description:
          type: string
          description: User description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample User document
        status:
          type: string
          description: User status
          enum:
            - active
            - archived
          example: active
    UserUpdate:
      type: object
      description: Fields to update in a User document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: User name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: User description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: User status
          enum:
            - active
            - archived
...